+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## v0.0.1 (09/26/22)
+## v0.0.2 (10/26/22)
+
+- Added support for the `Note` severity type, that Code scanning supports
+- Fixed the `View Alerts in GitHub` command in the `Security` view's toolbar
+- Code scanning alerts now display the tool that discovered them (e.g. CodeQL, ESLint)
+
+## v0.0.1 (10/20/22)
 
 - Initial release! 🚀